apps/trading-bot/requirements/base.txt
# Base requirements for Trading Bot
# =================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Financial Data
yfinance==0.2.28
ib-insync==0.9.86

# Sentiment Analysis
tweepy==4.14.0
vaderSentiment==3.3.2
praw==7.7.1
feedparser==6.0.10
newspaper3k==0.2.8
pytrends==4.9.2
beautifulsoup4==4.12.2
lxml==4.9.3

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
structlog==23.2.0
tenacity==8.2.3
psutil==5.9.6

# HTTP Client
httpx==0.25.2
sec-edgar-downloader==5.1.3
aiohttp==3.9.1
requests==2.31.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Monitoring & Metrics
prometheus-client==0.19.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
